+++ src/main/java/com/example/tileshop/repository/RoleRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Byte> {
+public interface RoleRepository extends JpaRepository<Role, Long> {
 
     Optional<Role> findByCode(RoleConstant code);
